ifupdownaddons: iproute2: fix set_mtu api to return if --no-act (DRYRUN) is set
Ticket: CM-10965
Reviewed By: julien
Testing Done: Tested set mtu with dry run

+++ ifupdownaddons/iproute2.py
@@ -419,8 +419,9 @@ class iproute2(utilsBase):
         self._cache_update([ifacename, 'hwaddress'], hwaddress)
 
     def link_set_mtu(self, ifacename, mtu):
+        if ifupdownflags.flags.DRYRUN:
+            return True
         if not mtu or not ifacename: return
-
         with open('/sys/class/net/%s/mtu' % ifacename, 'w') as f:
             f.write(mtu)
         self._cache_update([ifacename, 'mtu'], mtu)
